apps: move MultiClusterError type to its own file
